kernel - Change time_second to time_uptime for all expiration calculations
* Vet the entire kernel and change use cases for expiration calculations
  using time_second to use time_uptime instead.

* Protects these expiration calculations from step changes in the wall time,
  particularly needed for route table entries.

* Probably requires further variable type adjustments but the use of
  time_uptime instead if time_second is highly unlikely to ever overrun
  any demotions to int still present.

+++ sys/platform/pc64/isa/clock.c
@@ -926,7 +926,7 @@ inittodr(time_t base)
 
 	sec += tz.tz_minuteswest * 60 + (wall_cmos_clock ? adjkerntz : 0);
 
-	y = time_second - sec;
+	y = (int)(time_second - sec);
 	if (y <= -2 || y >= 2) {
 		/* badly off, adjust it */
 		ts.tv_sec = sec;
